Change Backend API
- Now the sotre method needs array of messages
- It's easier to handle blocks of message to optimize delivery
