+++ package.json
@@ -37,6 +37,7 @@
     "koa": "^2.0.0",
     "lodash": "^4.15.0",
     "mz": "^2.4.0",
-    "performance-now": "^0.2.0"
+    "performance-now": "^0.2.0",
+    "superagent": "^2.2.0"
   }
 }
